--- tvx/chain-validation/drivers/constants.go
@@ -1,15 +1,0 @@
-package drivers
-
-import (
-	big_spec "github.com/filecoin-project/specs-actors/actors/abi/big"
-)
-
-const (
-	totalFilecoin     = 2_000_000_000
-	filecoinPrecision = 1_000_000_000_000_000_000
-)
-
-var (
-	TotalNetworkBalance = big_spec.Mul(big_spec.NewInt(totalFilecoin), big_spec.NewInt(filecoinPrecision))
-	EmptyReturnValue    = []byte{}
-)
